LifeBalanceTracker_Project/requirement.txt
asgiref==3.9.1
certifi==2025.8.3
distlib==0.4.0
Django==5.2.5
django-csp==4.0
django-filter==25.1
djangorestframework==3.16.1
filelock==3.19.1
mysql-connector-python==9.4.0
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.4.0
setuptools==80.9.0
sqlparse==0.5.3
virtualenv==20.34.0
